requirements.txt
python-dotenv==1.0.0
aiohttp>=3.9.0
websockets>=13.0,<14.0

